+++ packages/observation-extractors/bitcoin-runes-observation-extractor/lib/bitcoinRunesEsploraObservationExtractor.ts
@@ -4,24 +4,29 @@ import { DataSource } from '@rosen-bridge/extended-typeorm';
 import { BitcoinRunesEsploraRosenExtractor } from '@rosen-bridge/rosen-extractor';
 import { TokenMap } from '@rosen-bridge/tokens';
 
+import { AbstractRunesProtocolNetwork } from './abstractRunesProtocolNetwork';
 import { BitcoinRunesAbstractObservationExtractor } from './bitcoinRunesAbstractObservationExtractor';
 
 export class BitcoinRunesEsploraObservationExtractor extends BitcoinRunesAbstractObservationExtractor<BitcoinEsploraTransaction> {
   constructor(
     lockAddress: string,
-    unisatUrl: string,
-    unisatApiKey: string,
+    runesProtocolNetwork: AbstractRunesProtocolNetwork,
     dataSource: DataSource,
     tokens: TokenMap,
     logger?: AbstractLogger,
+    storeRawData = true,
   ) {
     super(
       lockAddress,
-      unisatUrl,
-      unisatApiKey,
+      runesProtocolNetwork,
       dataSource,
       tokens,
-      new BitcoinRunesEsploraRosenExtractor(lockAddress, tokens, logger),
+      new BitcoinRunesEsploraRosenExtractor(
+        lockAddress,
+        tokens,
+        logger,
+        storeRawData,
+      ),
       logger,
     );
   }