+++ packages/abstract-observation-extractor/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @rosen-bridge/abstract-observation-extractor
 
+## 0.2.2
+
+### Patch Changes
+
+- Update Dependencies
+  - @rosen-bridge/rosen-extractor@11.1.0
+
+## 0.2.1
+
+### Patch Changes
+
+- Fixed type definitions in entities based migrations
+- Update Dependencies
+  - @rosen-bridge/rosen-extractor@11.0.0
+
 ## 0.2.0
 
 ### Minor Changes