+++ packages/observation-extractors/bitcoin-observation-extractor/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @rosen-bridge/bitcoin-observation-extractor
 
+## 6.4.0
+
+### Minor Changes
+
+- Add `storeRawData` option to store observations raw-data into the database
+
+### Patch Changes
+
+- Update Dependencies
+  - @rosen-bridge/abstract-observation-extractor@0.2.2
+  - @rosen-bridge/rosen-extractor@11.1.0
+
+## 6.3.1
+
+### Patch Changes
+
+- Update Dependencies
+  - @rosen-bridge/abstract-observation-extractor@0.2.1
+  - @rosen-bridge/bitcoin-scanner@0.2.2
+  - @rosen-bridge/rosen-extractor@11.0.0
+
 ## 6.3.0
 
 ### Minor Changes